Updated image_keywords_edinburgh.csv to be complete
This took fucking ages, and used up my year's worth of AWS Rekognition credits in one sitting. And, on top of that, I had to limit each listing to the main picture and no more than two others, as each image took around three seconds to produce labels, and there are around 7.5k listings for Edinburgh.